Update README.md
Added a few details to help with working through the examples. For instance, line 69 threw an error for me because I had not loaded the flights data frame with the data(flights) command.  I also added some explanation on lines 89 and 93 for what the nycflights13_sqlite() and nycflights13_postgresql() functions do.  Rearranged discussion of performance (lined 118 - 143) so that the examples track the description more exactly. 

Note - it would be nice to have a little more explanation of plyr and dplyr compatibility. Since they overlap in many functions and and the dplyr functions are faster and more general, what would be an example of when it would be a good idea to load both?